requirements/prod.txt
alembic==1.8.1
attrs==22.1.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
black==22.8.0
blinker==1.5
Bootstrap-Flask==2.1.0
cachelib==0.9.0
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
dnspython==2.2.1
email-validator==1.3.0
environs==9.5.0
factory-boy==3.2.1
Faker==15.0.0
fancycompleter==0.9.1
flake8==5.0.4
flake8-blind-except==0.2.1
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-isort==4.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.1
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.0.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.0
Flask-Static-Digest==0.2.1
Flask-WTF==1.0.1
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
mccabe==0.7.0
mypy-extensions==0.4.3
onetimepass==1.0.1
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
pdbpp==0.10.3
pep8-naming==0.13.2
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.1.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
PyQRCode==1.2.1
pyrepl==0.9.0
pytest==7.1.3
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
supervisor==4.2.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.2.2
wmctrl==0.4
WTForms==3.0.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
